% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_tables.R
\name{bind_tables}
\alias{bind_tables}
\title{This function combines all the raw data tables that contains raw data in one
year and remove all rows with NA in the table}
\usage{
bind_tables(t1, t2, t3, t4)
}
\arguments{
\item{t1}{The first raw data table to be combined}

\item{t2}{The second raw data table to be combined}

\item{t3}{The third raw data table to be combined}

\item{t4}{The fourth raw data table to be combined}
}
\value{
a data frame that contains all raw data in one year
}
\description{
This function combines all the raw data tables that contains raw data in one
year and remove all rows with NA in the table
}
\examples{
# bind_tables(table0007,table0107,table0207, table0307)
# bind_tables(table0008,table0108,table0208, table0308)

}
