% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_regression.R
\name{fit_regression}
\alias{fit_regression}
\title{The function accepts the wrangled Youtube dataset of type dataframe and fits a linear regression
model to the data.}
\usage{
fit_regression(traindata)
}
\arguments{
\item{traindata:}{Typically a portion of the dataset that is used for training a regression model}
}
\value{
a linear regression model with 'views' as the response variable and the chosen variables
as the explanatory variables
}
\description{
The function accepts the wrangled Youtube dataset of type dataframe and fits a linear regression
model to the data.
}
\examples{
# fit_regression(training)
}
