% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_uncleaned_data.R
\name{read_uncleaned_data}
\alias{read_uncleaned_data}
\title{This function takes in a path to where the uncleaned data file saved by load.R
located and read the data}
\usage{
read_uncleaned_data(path_to_file)
}
\arguments{
\item{path_to_file}{The path to where the file is located}
}
\value{
a data frame that contains uncleaned data saved by R/load.R
}
\description{
This function takes in a path to where the uncleaned data file saved by load.R
located and read the data
}
\examples{
# read_raw_data("/home/rstudio/data/data2007_not_cleaned.txt")
# read_raw_data("/home/rstudio/data/data2008_not_cleaned.txt")

}
