% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling_data.R
\name{wrangling_data}
\alias{wrangling_data}
\title{This function takes in a YouTube Data that read from raw data and
Remove unnecessary data and convert category variable as factor class}
\usage{
wrangling_data(data)
}
\arguments{
\item{data}{The YouTube data frame that reads from raw file and requires
removing it's unnecessary data and converting it's category variable}
}
\value{
a YouTube data frame that only contains "age", 'category',
'length','views','rate','ratings','comments' as columns; If it is
not a data frame, an error message would be "Please input a valid
data frame!";
}
\description{
This function takes in a YouTube Data that read from raw data and
Remove unnecessary data and convert category variable as factor class
}
\examples{
# wrangling_data(data2007_test)
# wrangling_data(data2008_test)

}
