% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_raw_data.R
\name{read_raw_data}
\alias{read_raw_data}
\title{This function takes in a url to where the file located and read the data}
\usage{
read_raw_data(url)
}
\arguments{
\item{url}{The URL to where the file is located}
}
\value{
a data frame that contains raw data in the file
}
\description{
This function takes in a url to where the file located and read the data
}
\examples{
data0007 <- read_raw_data("https://raw.githubusercontent.com/chris-ckh/dsci-310-group-13/main/data/0007.txt")
data0107 <- read_raw_data("https://raw.githubusercontent.com/chris-ckh/dsci-310-group-13/main/data/0107.txt")

}
